refactor: organize category widgets & deduplicate code
