+++ lib/features/categories/ui/widgets/category_selection_icon_button.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/features/categories/ui/widgets/category_color_icon.dart';
+import 'package:lotti/features/categories/ui/widgets/category_selection_modal_content.dart';
+import 'package:lotti/features/journal/state/entry_controller.dart';
+import 'package:lotti/l10n/app_localizations_context.dart';
+import 'package:lotti/utils/modals.dart';
+
+class CategorySelectionIconButton extends ConsumerWidget {
+  const CategorySelectionIconButton({
+    required this.entry,
+    super.key,
+  });
+
+  final JournalEntity entry;
+
+  @override
+  Widget build(
+    BuildContext context,
+    WidgetRef ref,
+  ) {
+    final provider = entryControllerProvider(id: entry.id);
+    final notifier = ref.read(provider.notifier);
+
+    return GestureDetector(
+      onTap: () {
+        ModalUtils.showSinglePageModal<void>(
+          context: context,
+          title: context.messages.habitCategoryLabel,
+          builder: (BuildContext _) {
+            return CategorySelectionModalContent(
+              onCategorySelected: (category) {
+                notifier.updateCategoryId(category?.id);
+                Navigator.pop(context);
+              },
+              initialCategoryId: entry.categoryId,
+            );
+          },
+        );
+      },
+      child: CategoryColorIcon(entry.categoryId),
+    );
+  }
+}
