+++ lib/features/habits/ui/widgets/habit_completion_color_icon.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
+import 'package:lotti/features/categories/ui/widgets/category_color_icon.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/entities_cache_service.dart';
