+++ lib/pages/settings/categories/categories_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
-import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
+import 'package:lotti/features/categories/ui/widgets/category_type_card.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/pages/settings/definitions_list_page.dart';
@@ -21,7 +21,7 @@ class CategoriesPage extends StatelessWidget {
       title: context.messages.settingsCategoriesTitle,
       getName: (category) => category.name.toLowerCase(),
       definitionCard: (int index, CategoryDefinition categoryDefinition) =>
-          CategoriesTypeCard(categoryDefinition, index: index),
+          CategoryTypeNavCard(categoryDefinition, index: index),
     );
   }
 }
